+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to `knotx-template-engine` will be documented in this file.
 
 ## Unreleased
 List of changes that are finished but not yet released in any final version.
-
+- [PR-21](https://github.com/Knotx/knotx-template-engine/pull/21) - Pebble Template Engine
 ## 2.1.0
 - [PR-18](https://github.com/Knotx/knotx-template-engine/pull/18) - `com.github.jknack` Handlebars updated to `4.1.2`.